Test: 게임 동기화 테스트를 위한 커밋
